+++ Tasks/src/main/java/streams/Main.java
@@ -1,5 +1,8 @@
 package streams;
 
+import streams.runners.ObjectiveOneRunner;
+import streams.runners.ObjectiveTwoRunner;
+
 public class Main {
 
     public static void main(String[] args) {
